# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Surah {
  id: number;
  nameSimple: string;
  nameArabic: string;
  nameComplex: string;
  versesCount: number;
  revelationPlace: string;
  translatedName: string;
}

export interface WordByWord {
  position: number;
  textUthmani: string;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  transliteration?: string | null;
}

export interface Verse {
  id: number;
  verseNumber: number;
  verseKey: string;
  surahNumber: number;
  textUthmani: string;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  wordByWord?: WordByWord[];
}

export interface VersesPage {
  verses: Verse[];
  totalCount: number;
  currentPage: number;
  totalPages: number;
}

export interface SearchResult {
  verseKey: string;
  surahName: string;
  verseNumber: number;
  textUthmani: string;
  /** @nullable */
  translation?: string | null;
}

export interface SearchResults {
  results: SearchResult[];
  totalCount: number;
  currentPage: number;
}

export interface Translation {
  id: number;
  name: string;
  authorName: string;
  languageName: string;
}

export interface Tafsir {
  id: number;
  name: string;
  authorName: string;
  languageName: string;
}

export interface TafsirContent {
  verseKey: string;
  tafsirId: number;
  text: string;
}

export interface Bookmark {
  id: number;
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface CreateBookmarkBody {
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface BookmarkStatus {
  bookmarked: boolean;
  /** @nullable */
  bookmarkId?: number | null;
}

export interface Note {
  id: number;
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  content: string;
}

export interface UpdateNoteBody {
  content: string;
}

export interface Goal {
  id: number;
  dailyVerseTarget: number;
  currentStreak: number;
  longestStreak: number;
  totalVersesRead: number;
  totalDaysRead: number;
  /** @nullable */
  lastReadAt?: string | null;
  createdAt: string;
}

export interface UpsertGoalBody {
  dailyVerseTarget: number;
}

export interface LogReadingSessionBody {
  versesRead: number;
}

export interface DayActivity {
  date: string;
  versesRead: number;
  read: boolean;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  totalDaysRead: number;
  /** @nullable */
  lastReadAt?: string | null;
  recentDays: DayActivity[];
}

export interface CollectionVerse {
  id: number;
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
  addedAt: string;
}

export interface Collection {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  verseCount: number;
  verses?: CollectionVerse[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateCollectionBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateCollectionBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface AddVerseToCollectionBody {
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
}

export interface DashboardSummary {
  totalBookmarks: number;
  totalNotes: number;
  totalCollections: number;
  currentStreak: number;
  longestStreak: number;
  totalVersesRead: number;
  dailyVerseTarget: number;
  /** @nullable */
  lastReadAt?: string | null;
  recentBookmarks: Bookmark[];
  recentNotes: Note[];
}

export type GetSurahVersesParams = {
  translationId?: number;
  page?: number;
  perPage?: number;
};

export type GetVerseParams = {
  translationId?: number;
};

export type SearchQuranParams = {
  q: string;
  translationId?: number;
  page?: number;
};

export type GetVerseTafsirParams = {
  tafsirId?: number;
};

export type GetRandomVerseParams = {
  translationId?: number;
};