# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Companion API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: Get all bookmarks
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkBody"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bookmarks/check/{verseKey}:
    get:
      operationId: checkBookmark
      tags: [bookmarks]
      summary: Check if a verse is bookmarked
      parameters:
        - name: verseKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmark status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"

  # ─── Notes / Reflections ────────────────────────────────────────────────────
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: Get all notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note for a verse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Goals & Streaks ────────────────────────────────────────────────────────
  /goals:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get user reading goal
      responses:
        "200":
          description: User reading goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    post:
      operationId: upsertGoal
      tags: [goals]
      summary: Create or update the reading goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertGoalBody"
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/log:
    post:
      operationId: logReadingSession
      tags: [goals]
      summary: Log a reading session (increments streak)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogReadingSessionBody"
      responses:
        "200":
          description: Updated goal with streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a collection with its verses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection with verses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionBody"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /collections/{id}/verses:
    post:
      operationId: addVerseToCollection
      tags: [collections]
      summary: Add a verse to a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVerseToCollectionBody"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}/verses/{verseKey}:
    delete:
      operationId: removeVerseFromCollection
      tags: [collections]
      summary: Remove a verse from a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: verseKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Quran ────────────────────────────────────────────────────
...
    Bookmark:
      type: object
      properties:
        id:
          type: integer
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        textUthmani:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - verseKey
        - surahNumber
        - surahName
        - verseNumber
        - createdAt
    CreateBookmarkBody:
      type: object
      properties:
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        textUthmani:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
      required:
        - verseKey
        - surahNumber
        - surahName
        - verseNumber
    BookmarkStatus:
      type: object
      properties:
        bookmarked:
          type: boolean
        bookmarkId:
          type: ["integer", "null"]
      required:
        - bookmarked

    # ─── Notes ─────────────────────────────────────────────────────────────
    Note:
      type: object
      properties:
        id:
          type: integer
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - verseKey
        - surahNumber
        - surahName
        - verseNumber
        - content
        - createdAt
        - updatedAt
    CreateNoteBody:
      type: object
      properties:
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        content:
          type: string
      required:
        - verseKey
        - surahNumber
        - surahName
        - verseNumber
        - content
    UpdateNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    # ─── Goals ─────────────────────────────────────────────────────────────
    Goal:
      type: object
      properties:
        id:
          type: integer
        dailyVerseTarget:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        totalVersesRead:
          type: integer
        totalDaysRead:
          type: integer
        lastReadAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - dailyVerseTarget
        - currentStreak
        - longestStreak
        - totalVersesRead
        - totalDaysRead
        - createdAt
    UpsertGoalBody:
      type: object
      properties:
        dailyVerseTarget:
          type: integer
      required:
        - dailyVerseTarget
    LogReadingSessionBody:
      type: object
      properties:
        versesRead:
          type: integer
      required:
        - versesRead
...
    CollectionVerse:
      type: object
      properties:
        id:
          type: integer
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        textUthmani:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - verseKey
        - surahNumber
        - surahName
        - verseNumber
        - addedAt
    Collection:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        verseCount:
          type: integer
        verses:
          type: array
          items:
            $ref: "#/components/schemas/CollectionVerse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - verseCount
        - createdAt
        - updatedAt
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
    UpdateCollectionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    AddVerseToCollectionBody:
      type: object
      properties:
        verseKey:
          type: string
        surahNumber:
          type: integer
        surahName:
          type: string
        verseNumber:
          type: integer
        textUthmani:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
      required:
        - verseKey
        - surahNumber
        - surahName
        - verseNumber

    # ─── Dashboard ─────────────────────────────────────────────────────────
...