# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}