# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionVerse } from "./collectionVerse";

export interface Collection {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  verseCount: number;
  verses?: CollectionVerse[];
  createdAt: Date;
  updatedAt: Date;
}