# lib/api-zod/src/generated/types/createBookmarkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookmarkBody {
  verseKey: string;
  surahNumber: number;
  surahName: string;
  verseNumber: number;
  /** @nullable */
  textUthmani?: string | null;
  /** @nullable */
  translation?: string | null;
  /** @nullable */
  note?: string | null;
}