# lib/api-zod/src/generated/types/updateCollectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Companion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCollectionBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}